Move FFT1D to kernels::signal
